Update PDFRecon.py
Added the ability to compare revisions to the original to see what has been chaged.
How the New Feature Works
New Dependency: The script now requires the Pillow library. If it's not installed, the program will show an error message on startup. You can install it by running: pip install Pillow.

Triggering the Comparison: In the results table, if you right-click on a row that is a revision (e.g., "Revision of #1"), a new option will appear in the context menu: "Visually Compare Revision".

Visual Diff Window: Clicking this option will open a new window that displays the first page of the Original document, the Revision, and a third image called Differences. This "diff" image will highlight the parts of the page that have been changed in red, making alterations easy to spot.